% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_goodbye.R
\name{say_goodbye}
\alias{say_goodbye}
\title{Say good bye to someone}
\usage{
say_goodbye(someone = "You")
}
\arguments{
\item{someone}{Character.}
}
\value{
Side effect: A message in the console
}
\description{
Say good bye to someone
}
\examples{
say_goodbye() #Good bye You !
say_goodbye(someone = "Christine") #Good bye Christine !
}
