% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Say hello to someone}
\usage{
say_hello(someone = "World")
}
\arguments{
\item{someone}{Character.}
}
\value{
Side effect: A message in the console
}
\description{
Say hello to someone
}
\examples{
say_hello() #Hello World
say_hello(someone = "Christine") #Hello Christine
}
